backend/requirements.txt
fastapi
uvicorn
python-multipart
langgraph
langchain-google-genai
pandas
SpeechRecognition
gTTS
python-dotenv
pydub